Feature: Announcements Section Feature: Saved information Via SessionStorage (TEMP) Feature: Added Assignments Section Feature: Added Name for input to get actual name due to checking numbers
I physically cannot fix the bug where you get the same comments for each classroom announcement because I have to save each info to a db -> set a table for the classroom -> set a column for each announcement -> containing each announcement... There's no way I can do that right now but I can  update it in the future. I know, it's not ideal.